requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-httplib2>=0.1.1
requests>=2.31.0
PyJWT[crypto]>=2.8.0
cryptography>=41.0.0
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.7.0
